Added 'setupcmd' subcommand and Markdown manual.
'setupcmd' prints the nest command that will recreate its
configuration so I don't have to keep trying to remember the magic
incantion to set up a new account.

The Markdown manual is just the POD section converted to Markdown with
pod2markdown, but now you can read it on Github.